Moves getCityData and formatCityData to City model - still works
